unsave content reworked

+++ routes/api/saveItems.js
@@ -3,6 +3,7 @@ const express = require("express");
 const requestValidator = require("../../middlewares/requestValidator");
 
 const SavedItem = require("../../models/SavedItem");
+const validateObjectId = require("../../helpers/validateObjectId");
 
 const {
   saveItemSchema,
@@ -42,13 +43,18 @@ router.post(
 );
 
 router.delete(
-  "/unsave_content",
+  "/unsave_content/:content",
   requestValidator(unsaveItemSchema),
   authorize(),
   async (req, res) => {
-    const { content } = _.pick(req.body, ["content"]);
-    const { user } = req.authSession;
+    const { content } = req.params;
 
+    if (!validateObjectId(content))
+      return res
+        .status(404)
+        .send({ error: { message: "Saved Content not found!" } });
+
+    const { user } = req.authSession;
     const savedItem = await SavedItem.findOneAndRemove({
       content,
       savedBy: user._id,
